+++ packages/api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @secondlayer/api
 
+## 1.22.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @secondlayer/shared@6.33.3
+  - @secondlayer/platform@0.1.6
+
 ## 1.22.6
 
 ### Patch Changes