+++ packages/platform/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @secondlayer/platform
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @secondlayer/shared@6.33.3
+
 ## 0.1.5
 
 ### Patch Changes